fix update upload buku dan gambar

+++ application/controllers/Data.php
@@ -8,7 +8,7 @@ class Data extends CI_Controller {
      $this->data['CI'] =& get_instance();
      $this->load->helper(array('form', 'url'));
      $this->load->model('M_Admin');
-		if($this->session->userdata('masuk_sistem_rekam') != TRUE){
+		if($this->session->userdata('masuk_perpus') != TRUE){
 				$url=base_url('login');
 				redirect($url);
 		}
@@ -87,18 +87,24 @@ class Data extends CI_Controller {
 
 	public function prosesbuku()
 	{
+		if($this->session->userdata('masuk_perpus') != TRUE){
+			$url=base_url('login');
+			redirect($url);
+		}
+
+		// hapus aksi form proses buku
 		if(!empty($this->input->get('buku_id')))
 		{
         
 			$buku = $this->M_Admin->get_tableid_edit('tbl_buku','id_buku',htmlentities($this->input->get('buku_id')));
 			
-			$sampul = './assets_style/image/buku/'.$buku->sampul;
+			$sampul = './assets/image/buku/'.$buku->sampul;
 			if(file_exists($sampul))
 			{
 				unlink($sampul);
 			}
 			
-			$lampiran = './assets_style/image/buku/'.$buku->lampiran;
+			$lampiran = './assets/image/buku/'.$buku->lampiran;
 			if(file_exists($lampiran))
 			{
 				unlink($lampiran);
@@ -107,133 +113,74 @@ class Data extends CI_Controller {
 			$this->M_Admin->delete_table('tbl_buku','id_buku',$this->input->get('buku_id'));
 			
 			$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-warning">
-			<p> Berhasil Hapus Buku !</p>
-			</div></div>');
+					<p> Berhasil Hapus Buku !</p>
+				</div></div>');
 			redirect(base_url('data'));  
 		}
+
+		// tambah aksi form proses buku
 		if(!empty($this->input->post('tambah')))
 		{
-
 			$post= $this->input->post();
-			// setting konfigurasi upload
-			$config['upload_path'] = './assets_style/image/buku/';
-			$config['allowed_types'] = 'gif|jpg|jpeg|png|pdf|doc'; 
-			$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
-			// load library upload
-			$this->load->library('upload',$config);
 			$buku_id = $this->M_Admin->buat_kode('tbl_buku','BK','id_buku','ORDER BY id_buku DESC LIMIT 1'); 
+			$data = array(
+				'buku_id'=>$buku_id,
+				'id_kategori'=>htmlentities($post['kategori']), 
+				'id_rak' => htmlentities($post['rak']), 
+				'isbn' => htmlentities($post['isbn']), 
+				'title'  => htmlentities($post['title']), 
+				'pengarang'=> htmlentities($post['pengarang']), 
+				'penerbit'=> htmlentities($post['penerbit']),    
+				'thn_buku' => htmlentities($post['thn']), 
+				'isi' => $this->input->post('ket'), 
+				'jml'=> htmlentities($post['jml']),  
+				'tgl_masuk' => date('Y-m-d H:i:s')
+			);
 
-			// upload gambar 1
-			if(!empty($_FILES['gambar']['name'] && $_FILES['lampiran']['name']))
+			$this->load->library('upload',$config);
+			if(!empty($_FILES['gambar']['name']))
 			{
-
+				// setting konfigurasi upload
+				$config['upload_path'] = './assets_style/image/buku/';
+				$config['allowed_types'] = 'gif|jpg|jpeg|png'; 
+				$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
+				// load library upload
+				$this->load->library('upload',$config);
 				$this->upload->initialize($config);
 
 				if ($this->upload->do_upload('gambar')) {
 					$this->upload->data();
 					$file1 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
+					$this->db->set('sampul', $file1['upload_data']['file_name']);
+				}else{
+					$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-success">
+							<p> Edit Buku Gagal !</p>
+						</div></div>');
+					redirect(base_url('data')); 
 				}
+			}
 
+			if(!empty($_FILES['lampiran']['name']))
+			{
+				// setting konfigurasi upload
+				$config['upload_path'] = './assets_style/image/buku/';
+				$config['allowed_types'] = 'pdf'; 
+				$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
+				// load library upload
+				$this->load->library('upload',$config);
+				$this->upload->initialize($config);
 				// script uplaod file kedua
 				if ($this->upload->do_upload('lampiran')) {
 					$this->upload->data();
 					$file2 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
+					$this->db->set('lampiran', $file2['upload_data']['file_name']);
+				}else{
+
+					$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-success">
+							<p> Edit Buku Gagal !</p>
+						</div></div>');
+					redirect(base_url('data')); 
 				}
-				$data = array(
-					'buku_id'=>$buku_id,
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => $file1['upload_data']['file_name'],
-                    'lampiran' => $file2['upload_data']['file_name'],
-					'title'  => htmlentities($post['title']), 
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-				
-
-			}elseif(!empty($_FILES['gambar']['name'])){
-				$this->upload->initialize($config);
-
-				if ($this->upload->do_upload('gambar')) {
-					$this->upload->data();
-					$file1 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
-				}
-				$data = array(
-					'buku_id'=>$buku_id,
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => $file1['upload_data']['file_name'],
-                    'lampiran' => '0',
-					'title'  => htmlentities($post['title']), 
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-			}elseif(!empty($_FILES['lampiran']['name'])){
-
-				$this->upload->initialize($config);
-
-				// script uplaod file kedua
-				if ($this->upload->do_upload('lampiran')) {
-					$this->upload->data();
-					$file2 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
-				}
-
-				// script uplaod file kedua
-				$this->upload->do_upload('lampiran');
-				$file2 = array('upload_data' => $this->upload->data());
-				$data = array(
-					'buku_id'=>$buku_id,
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => '0',
-                    'lampiran' => $file2['upload_data']['file_name'],
-					'title'  => htmlentities($post['title']), 
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-				
-			}else{
-				$data = array(
-					'buku_id'=>$buku_id,
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => '0',
-                    'lampiran' => '0',
-					'title'  => htmlentities($post['title']), 
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),    
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
 			}
 
 			$this->db->insert('tbl_buku', $data);
@@ -244,157 +191,84 @@ class Data extends CI_Controller {
 			redirect(base_url('data')); 
 		}
 
+		// edit aksi form proses buku
 		if(!empty($this->input->post('edit')))
 		{
-			$post= $this->input->post();
-			// setting konfigurasi upload
-			$config['upload_path'] = './assets_style/image/buku/';
-			$config['allowed_types'] = 'gif|jpg|jpeg|png'; 
-			$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
-			// load library upload
-        	$this->load->library('upload',$config);
-			// upload gambar 1
-			if(!empty($_FILES['gambar']['name'] && $_FILES['lampiran']['name']))
+			$post = $this->input->post();
+			$data = array(
+				'id_kategori'=>htmlentities($post['kategori']), 
+				'id_rak' => htmlentities($post['rak']), 
+				'isbn' => htmlentities($post['isbn']), 
+				'title'  => htmlentities($post['title']),
+				'pengarang'=> htmlentities($post['pengarang']), 
+				'penerbit'=> htmlentities($post['penerbit']),  
+				'thn_buku' => htmlentities($post['thn']), 
+				'isi' => $this->input->post('ket'), 
+				'jml'=> htmlentities($post['jml']),  
+				'tgl_masuk' => date('Y-m-d H:i:s')
+			);
+
+			if(!empty($_FILES['gambar']['name']))
 			{
-
+				// setting konfigurasi upload
+				$config['upload_path'] = './assets_style/image/buku/';
+				$config['allowed_types'] = 'gif|jpg|jpeg|png'; 
+				$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
+				// load library upload
+				$this->load->library('upload',$config);
 				$this->upload->initialize($config);
 
 				if ($this->upload->do_upload('gambar')) {
 					$this->upload->data();
+					$gambar = './assets/image/buku/'.htmlentities($post['gmbr']);
+					if(file_exists($gambar)) {
+						unlink($gambar);
+					}
 					$file1 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
+					$this->db->set('sampul', $file1['upload_data']['file_name']);
+				}else{
+					$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-success">
+							<p> Edit Buku Gagal !</p>
+						</div></div>');
+					redirect(base_url('data')); 
 				}
+			}
 
+			if(!empty($_FILES['lampiran']['name']))
+			{
+				// setting konfigurasi upload
+				$config['upload_path'] = './assets_style/image/buku/';
+				$config['allowed_types'] = 'pdf'; 
+				$config['encrypt_name'] = TRUE; //nama yang terupload nantinya
+				// load library upload
+				$this->load->library('upload',$config);
+				$this->upload->initialize($config);
 				// script uplaod file kedua
 				if ($this->upload->do_upload('lampiran')) {
 					$this->upload->data();
+					$lampiran = './assets_style/image/buku/'.htmlentities($post['lamp']);
+					if(file_exists($lampiran)) {
+						unlink($lampiran);
+					}
 					$file2 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
+					$this->db->set('lampiran', $file2['upload_data']['file_name']);
+				}else{
+
+					$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-success">
+							<p> Edit Buku Gagal !</p>
+						</div></div>');
+					redirect(base_url('data')); 
 				}
-
-				$gambar = './assets_style/image/buku/'.htmlentities($post['gmbr']);
-				if(file_exists($gambar))
-				{
-					unlink($gambar);
-				}
-
-				$lampiran = './assets_style/image/buku/'.htmlentities($post['lamp']);
-				if(file_exists($lampiran))
-				{
-					unlink($lampiran);
-				}
-
-				$data = array(
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => $file1['upload_data']['file_name'],
-                    'lampiran' => $file2['upload_data']['file_name'],
-					'title'  => htmlentities($post['title']),
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-				
-
-			}elseif(!empty($_FILES['gambar']['name'])){
-				$this->upload->initialize($config);
-
-				if ($this->upload->do_upload('gambar')) {
-					$this->upload->data();
-					$file1 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
-				}
-
-
-				$gambar = './assets_style/image/buku/'.htmlentities($post['gmbr']);
-				if(file_exists($gambar))
-				{
-					unlink($gambar);
-				}
-
-				$data = array(
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'sampul' => $file1['upload_data']['file_name'],
-					'title'  => htmlentities($post['title']),
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-			}elseif(!empty($_FILES['lampiran']['name'])){
-
-				$this->upload->initialize($config);
-
-				// script uplaod file kedua
-				if ($this->upload->do_upload('lampiran')) {
-					$this->upload->data();
-					$file2 = array('upload_data' => $this->upload->data());
-				} else {
-					return false;
-				}
-
-				$lampiran = './assets_style/image/buku/'.htmlentities($post['lamp']);
-				if(file_exists($lampiran))
-				{
-					unlink($lampiran);
-				}
-
-				// script uplaod file kedua
-				$this->upload->do_upload('lampiran');
-				$file2 = array('upload_data' => $this->upload->data());
-
-				$data = array(
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-                    'lampiran' => $file2['upload_data']['file_name'],
-					'title'  => htmlentities($post['title']),
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),  
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
-
-				
-			}else{
-				$data = array(
-					'id_kategori'=>htmlentities($post['kategori']), 
-					'id_rak' => htmlentities($post['rak']), 
-					'isbn' => htmlentities($post['isbn']), 
-					'title'  => htmlentities($post['title']), 
-					'pengarang'=> htmlentities($post['pengarang']), 
-					'penerbit'=> htmlentities($post['penerbit']),    
-					'thn_buku' => htmlentities($post['thn']), 
-					'isi' => $this->input->post('ket'), 
-					'jml'=> htmlentities($post['jml']),  
-					'tgl_masuk' => date('Y-m-d H:i:s')
-				);
 			}
 
 			$this->db->where('id_buku',htmlentities($post['edit']));
 			$this->db->update('tbl_buku', $data);
 
 			$this->session->set_flashdata('pesan','<div id="notifikasi"><div class="alert alert-success">
-			<p> Edit Buku Sukses !</p>
-			</div></div>');
-			redirect(base_url('data')); 
+					<p> Edit Buku Sukses !</p>
+				</div></div>');
+			redirect(base_url('data/bukuedit/'.$post['edit'])); 
 		}
-		
 	}
 
 	public function kategori()
